+++ lib/src/answer_format/boolean_answer_format.dart
@@ -12,7 +12,7 @@ class BooleanAnswerFormat implements AnswerFormat {
   const BooleanAnswerFormat({
     required this.positiveAnswer,
     required this.negativeAnswer,
-    this.result = BooleanResult.NONE,
+    this.result = BooleanResult.none,
   }) : super();
 
   factory BooleanAnswerFormat.fromJson(Map<String, dynamic> json) =>
@@ -21,4 +21,4 @@ class BooleanAnswerFormat implements AnswerFormat {
   Map<String, dynamic> toJson() => _$BooleanAnswerFormatToJson(this);
 }
 
-enum BooleanResult { NONE, POSITIVE, NEGATIVE }
+enum BooleanResult { none, positive, negative }