+++ lib/src/result/survey/survey_result.dart
@@ -17,8 +17,9 @@ class SurveyResult extends Result {
     required this.finishReason,
     required this.results,
   }) : super(id: id, startDate: startDate, endDate: endDate);
-  
-  factory SurveyResult.fromJson(Map<String, dynamic> json) => _$SurveyResultFromJson(json);
+
+  factory SurveyResult.fromJson(Map<String, dynamic> json) =>
+      _$SurveyResultFromJson(json);
 
   Map<String, dynamic> toJson() => _$SurveyResultToJson(this);
 
@@ -26,4 +27,4 @@ class SurveyResult extends Result {
   List<Object?> get props => [id, startDate, endDate, finishReason];
 }
 
-enum FinishReason { SAVED, DISCARDED, COMPLETED, FAILED }
+enum FinishReason { saved, discarded, completed, failed }