+++ lib/src/steps/predefined_steps/cms_step.dart
@@ -4,9 +4,10 @@ import 'package:survey_kit/src/steps/cms_steps/cms_content_step.dart';
 import 'package:survey_kit/src/views/cms_step_view.dart';
 
 @JsonSerializable()
+@immutable
 class CMSStep {
   final List<Content> contextSteps;
-  CMSStep({
+  const CMSStep({
     required this.contextSteps,
   });
 
@@ -17,12 +18,8 @@ class CMSStep {
   }
 
   @override
-  bool operator ==(Object o) => super == o && o is CMSStep;
+  bool operator ==(Object other) => super == other && other is CMSStep;
 
   @override
-  // TODO: implement hashCode
-  int get hashCode => super.hashCode;
-
-
-
+  int get hashCode => super.hashCode ^ contextSteps.hashCode;
 }